Reorg PlaybookTrigger code
All Playbook can have triggers, so move to base class
